this a nother hot fix

+++ 0-bubble_sort.c
@@ -7,7 +7,7 @@
  */
 void bubble_sort(int *array, size_t size)
 {
-	size_t i, j;
+	unsigned int i, j;
 	int temp;
 
 	if (size < 2)
